requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.7.4
cftime==1.6.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
dgl==2.2.1
einops==0.8.0
einops-exts==0.0.4
ema-pytorch==0.6.0
fastjsonschema==2.20.0
filelock==3.13.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.2.0
grpcio==1.66.0
huggingface-hub==0.24.6
idna==3.8
iniconfig==2.0.0
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.0.5
nbformat==5.10.4
netCDF4==1.7.1.post2
networkx==3.2.1
numpy==1.26.3
pandas==2.2.2
pillow==10.2.0
plotly==5.24.0
pluggy==1.5.0
protobuf==5.27.3
pyg_lib==0.4.0+pt24cu118
pyparsing==3.1.3
pytest==8.3.3
pytz==2024.1
pywin32==305.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rotary-embedding-torch==0.6.5
rpds-py==0.20.0
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
setuptools==72.1.0
sympy==1.12
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
timm==1.0.9
torch==2.4.0+cu118
torch_cluster==1.6.3+pt24cu118
torch_geometric==2.5.3
torch_scatter==2.1.2+pt24cu118
torch_sparse==0.6.18+pt24cu118
torch_spline_conv==1.2.2+pt24cu118
torchaudio==2.4.0+cu118
torchdata==0.8.0
torchvision==0.19.0+cu118
tqdm==4.66.5
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
wheel==0.43.0
yacs==0.1.8
yarl==1.9.4
